skydive::mldbg
Struct MLdBG

pub struct MLdBG {
    pub kmer_size: usize,
    pub ldbgs: Vec<LdBG>,
    pub scores: HashMap<Vec<u8>, f32>,
}

Represents a multi-color linked de Bruijn graph, all built with same k-mer size.

Fields
kmer_size: usizeldbgs: Vec<LdBG>scores: HashMap<Vec<u8>, f32>
Implementations

impl MLdBG

pub fn new(kmer_size: usize) -> Self

Create an empty multi-color LdBG.

pub fn from_ldbgs(ldbgs: Vec<LdBG>) -> Self

Create an MLdBG from a vector of LdBGs.

Arguments

ldbgs - A vector of LdBGs.

Returns

A new MLdBG.

Panics

This function will panic if the kmer_size of the LdBG being added does not match the kmer_size of the MLdBG. Specifically, it will panic at the assert! statement if the condition ldbg.kmer_size == self.kmer_size is not met.

pub fn push(&mut self, ldbg: LdBG)

Add a LdBG to the MLdBG.

Arguments

ldbg - The LdBG to add.

Panics

The push function will panic if the kmer_size of the ldbg being added does not match the kmer_size of the MLdBG. Specifically, it will panic at the assert! statement if the condition ldbg.kmer_size == self.kmer_size is not met.

pub fn insert(&mut self, index: usize, ldbg: LdBG)

Insert a LdBG at a specific position in the MLdBG.

pub fn append(&mut self, ldbg: LdBG)

Append a LdBG to the end of the MLdBG.

pub fn append_from_file(&mut self, name: String, seq_path: &PathBuf)

Append an LdBG to the end of the MLdBG, created anew from a fasta file.

pub fn append_from_filtered_file<F>( &mut self, name: String, seq_path: &PathBuf, filter: F, )
where F: Fn(&Record, &HashSet<Vec<u8>>) -> bool,

Append an LdBG to the end of the MLdBG, created from a filtered set of sequences in a fasta file.

This function reads sequences from a specified fasta file, filters them based on a provided condition, and then creates an LdBG from the filtered sequences. The new LdBG is appended to the MLdBG.

Arguments

name - A string representing the name of the new LdBG.

seq_path - A reference to a PathBuf representing the path to the fasta file containing the sequences.

filter - A closure that takes a fasta record and a set of kmers and returns a boolean indicating whether the record should be included.

Panics

This function will panic if it cannot read the fasta file.

pub fn score_kmers(self, model_path: &PathBuf) -> Self

Scores the k-mers in the MLdBG using a pre-trained Gradient Boosting Decision Tree (GBDT) model.

This function loads a GBDT model from the specified path and uses it to predict scores for each k-mer in the union of k-mers from all LdBGs in the MLdBG. The scores are stored in the scores field of the MLdBG.

Arguments

model_path - A path to the file containing the pre-trained GBDT model.

Returns

An updated MLdBG with the k-mer scores populated.

Panics

This function will panic if the model file cannot be loaded or if the prediction fails.

pub fn collapse(&mut self) -> LdBG

pub fn filter_reads<F>(&mut self, seq_path: &PathBuf, filter: F) -> Vec<Vec<u8>>
where F: Fn(&Record, &HashSet<Vec<u8>>) -> bool,

Filter reads from a fasta file based on a condition.

Arguments

seq_path - A path to the fasta file containing the reads.

filter - A closure that takes a fasta record and a set of kmers and returns a boolean.

Returns

A vector of the kept reads.

Panics

This function will panic if it cannot read the fasta file.

pub fn union_of_kmers(&self) -> HashSet<Vec<u8>>

Get the union of kmers from all LdBGs in the MLdBG.

pub fn get(&self, index: usize) -> Option<&LdBG>

Get a reference to the LdBG at a specific index.

pub fn iter(&self) -> Iter<'_, LdBG>

Returns an iterator over the LdBGs in the MLdBG.

pub fn iter_mut(&mut self) -> IterMut<'_, LdBG>

Returns a mutable iterator over the LdBGs in the MLdBG.

pub fn clear(&mut self)

Clear all LdBGs from the MLdBG.

pub fn remove(&mut self, index: usize) -> Option<LdBG>

Remove a LdBG from the MLdBG by index.

pub fn len(&self) -> usize

Returns the number of LdBGs in the MLdBG.

pub fn is_empty(&self) -> bool

Check if the MLdBG is empty.

pub fn pop(&mut self) -> Option<LdBG>

Remove and return the last LdBG from the MLdBG.

pub fn pop_if<F>(&mut self, condition: F) -> Option<LdBG>
where F: Fn(&LdBG) -> bool,

Remove and return the LdBG from the MLdBG if it matches a certain condition.