skydive::storage_local
Function local_get_file_update_time

pub fn local_get_file_update_time(path: &PathBuf) -> Result<DateTime<Utc>>

Get the update time of a file in the local filesystem.

Arguments

path - A reference to a PathBuf object representing the path to the file.

Returns

A DateTime<Utc> object representing the update time of the file.

Errors

This function returns an error if the file metadata cannot be accessed.

Panics

This function panics if the file path is invalid.