skydive

Module utils

Functions

basename_without_extension

This function takes a sequence URL and a list of possible extensions, and returns the base name of the file without any of the provided extensions. It does this by first extracting the last segment of the URL path, and then iteratively removing any of the specified extensions from the end of the base name.

canonicalize_kmer

Get the canonical (lexicographically-lowest) version of a k-mer.

default_bounded_progress_bar

Create a new bounded progress bar with the specified message and length. The progress bar will be a bar with a spinner. The progress bar will display the elapsed time, the progress bar, the current position, the total length, and the estimated time remaining.

default_hidden_progress_bar

default_unbounded_progress_bar

Create a new unbounded progress bar with the specified message.

gc_content

homopolymer_compressed

read_fasta

Given fasta files this function will read and return a list of lists containing the contents of the fasta files

read_gfa

Reads a GFA file and constructs a directed graph from it.

shannon_entropy

write_gfa

Writes a GFA file from a directed graph.