Function skydive::utils::basename_without_extension

pub fn basename_without_extension(seq_url: &Url, extensions: &[&str]) -> String

This function takes a sequence URL and a list of possible extensions, and returns the base name of the file without any of the provided extensions. It does this by first extracting the last segment of the URL path, and then iteratively removing any of the specified extensions from the end of the base name.

Arguments

seq_url - A reference to a URL object representing the sequence file URL.

extensions - A slice of string slices representing the possible file extensions to be removed.

Returns

A String containing the base name of the file without any of the specified extensions.

Example

let url = url::Url::parse("http://example.com/path/to/file.fasta.gz").unwrap();
let extensions = [".fasta.gz", ".fa.gz", ".fasta", ".fa"];
let basename = basename_without_extension(&url, &extensions);
assert_eq!(basename, "file");