skydive::utils
Function read_gfa

pub fn read_gfa<P: AsRef<Path>>(path: P) -> Result<DiGraph<String, f32>>

Reads a GFA file and constructs a directed graph from it.

Arguments

path - A path to the GFA file.

Returns

A DiGraph where nodes are sequences and edges are links with weights.

Errors

This function returns an error if the file cannot be opened or read.

Panics

This function will panic if:

The file cannot be opened.

Any line in the file cannot be read.